Add files via upload
The latest and the greatest in pairs trade modeling. Gains have been reduced, but the strategy is still A) Market Neutral B) Feasible.
Added in a line to calculate the number of trades over the lifetime of the deal, and it's a lot. The strategy is feasible for bots, but not people.

Fixed a scalar error that occurred when trying to use values about 0.5 for aggressiveness. 

We could run a test to determine the best parameters, but I found the decent results with 0.75 buy aggressiveness, 0.9 sell aggressiveness, and 1.65 zScore (10% significance). zScore does not make a huge impact anymore, as finding a profitable spread seems to outweigh it.